refactor: 코멘트 패키지 리팩토링
- domain -> entity로 패키지명 변경
- mapping 메서드 mapper클래스에서 수행하도록 변경
